- change ifdi_rx_intr_enable to take a 16 bit queue id - change vaddrs in iflib_qset_addr_get to caddr_t - add "friend" functions to give access to iflib's sctx and qset   mutexes so that drivers with special needs (e.g. cxgb which supports   upper level protocols) can synchronize with iflib

+++ sys/net/ifdi_if.m
@@ -58,7 +58,7 @@ CODE {
 	}
 
 	static void
-	null_queue_intr_enable(if_shared_ctx_t _ctx __unused, uint32_t _qid __unused)
+	null_queue_intr_enable(if_shared_ctx_t _ctx __unused, uint16_t _qid __unused)
 	{
 	}
 
@@ -146,7 +146,7 @@ METHOD void intr_disable {
 
 METHOD void rx_intr_enable {
 	if_shared_ctx_t _ctx;
-	uint32_t _rxqid;
+	uint16_t _rxqid;
 } DEFAULT null_queue_intr_enable;
 
 METHOD void link_intr_enable {
